% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot cases over state map for a given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Intger representing the state code of the state to be plotted.}

\item{year}{Integer representing the year of the cases to be plotted.}
}
\value{
This function returns a map with the cases in given year over the state.
}
\description{
This function takes a year and state code as arguments and plots the cases over the state map.
}
\details{
This function uses the \code{maps::map()} and \code{graphics::points()} functions.
}
