% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Read File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string containing the name of the csv data file to read in.}
}
\value{
This function returns a tibble containing the data from the filename provided.
}
\description{
This is a simple read function that reads in csv file for a given filename.
}
\details{
This function uses the \code{readr::read_csv()} and \code{dplyr::tbl_df()} functions.
}
\examples{
fars_read("accident_2015.csv.bz2")

}
