% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{List cases by month and year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector with years to be listed.}
}
\value{
This function returns a list of tibbles for each valid year provided in the input.
}
\description{
This helper function lists, by year, tibbles featuring the year and month of all cases of given year.
}
\details{
This function uses \code{dplyr}'s \code{mutate} and \code{select} functions.
}
