% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize count of cases per month by year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Vector of years to be summarized.}
}
\value{
This function returns a tibble featuring the counts of cases with Month as rows, and Year as columns.
}
\description{
This function summarizes the counts of cases per month for each year provided in the input.
}
\details{
This function uses \code{tidyr::spread()} and \code{dplyr}'s \code{group_by}, \code{bind_rows}, and \code{summarize} functions.
}
\examples{
fars_summarize_years(c(2014, 2015))

}
