% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{Get filename from year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer representing the year.}
}
\value{
This function returns a character string containing the filename corresponding to the year input.
}
\description{
This function outputs the filename corresponding to a given year.
}
